Optimize getting old commits
This is also fixing the problem that the commit itself was used as one
of the old commits.  This was causing contributor checks to miss some
inconsistencies.  Checking contributors became really slow as a side
effect of this fix as now it really needs to iterate through the old
commits until it finds the relevant ones.