Throw away content process piping optimization
I had initially designed the tool to pipe the Git content process
directly to the syntax checkers to avoid buffering the files on Python.
I tried to preserve this design while adding more features, but it got
complicated.  This complexity caused file descriptors to leak, errors
to be silently ignored.  I though I had fixed them, but as reported
on #7 I was not.

After doing some performance testing, this optimization turn out to be
not worthwhile.  The tool would need more memory, and can get a little
slower while there are very big files on Git, but I don't worry about it
anymore.  Git is not particularly fast at dealing with big files anyway.
It doesn't make any difference on normal cases.  It can even make things
faster when there are multiple syntax checkers for a single file as we
are caching the contents now.  It would also allow us to develop more
features inside the tool as we have free access to the file content now.

This should fix #7.  I don't have access to the repository the problem
is reported, but I am able to reproduce similar problems without this
commit.