Support tags
This provides a basic support for the Git tags.  The previous version
was just crashing when a tag is pushed.  This now skips the checks on
list of commits, and applies the checks for single commits to the tags.
Surprisingly, the checks for single commits work on the tags, and
the checks for the committed files don't apply.

We would probably improve the support for tags.  We can check perhaps
relate them with the appropriate list of commits, and cross check
properties like the contributors.